% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attitude.R
\name{attitude}
\alias{attitude}
\title{attitude}
\usage{
attitude(erect = 0, ascending = 45, spreading = 90, reflexed = 135,
  ylim = c(0, 100), ...)
}
\arguments{
\item{erect}{degrees for erect structures.}

\item{ascending}{degrees for ascending structures.}

\item{spreading}{degrees for spreading structures.}

\item{reflexed}{degrees for reflexed structures.}

\item{ylim}{the y limits of the plot.}

\item{...}{further parameters passed to other methods.}
}
\description{
Plot attitude of structures
}
\details{
Plot attitude of structures
}
\examples{
attitude()

}

