% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_report.R
\name{character_report}
\alias{character_report}
\title{character_report}
\usage{
character_report(chars, file = "character_report.Rmd", meta = TRUE)
}
\arguments{
\item{chars}{a data.frame containing samples (rows) and characters (columns).}

\item{file}{a character string giving the output file name.}

\item{meta}{logical specifying whether the meta data should be printed.}
}
\description{
Prepare a report for characters
}
\details{
Prepare an RMarkdown report for characters.
}
\examples{
data("iris")
character_report( iris[,-5] )

}

