% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/character_report2.R
\name{character_report2}
\alias{character_report2}
\title{character_report2}
\usage{
character_report2(x)
}
\arguments{
\item{x}{a data.frame containing samples (rows) and characters (columns).}
}
\description{
Create a RMarkdown report about character classes and distribution
}
\details{
Prepare an RMarkdown report for characters.
}
\examples{
myData <- data.frame(matrix(ncol=4, nrow=12))
colnames(myData) <- paste('char', 1:ncol(myData), sep="_")
rownames(myData) <- paste('sample', 1:nrow(myData), sep="_")
set.seed(9)
myData[,1] <- abs(round(rnorm(nrow(myData), mean = 5, sd =2)))
myData[,2] <- letters[myData[,1]]
myData[,3] <- as.factor(myData[,2])
myData[,4] <- factor(myData[,2], ordered = TRUE)
myData[3,2] <- "?" # Add an nexpected value

\dontrun{
character_report2(myData)
}
}

