% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDF.R
\name{cleanDF}
\alias{cleanDF}
\title{cleanDF}
\usage{
cleanDF(df)
}
\arguments{
\item{df}{an object of class data.frame}
}
\description{
Sanitizes objects of class data.frame
}
\details{
Removes leading and trailing whitespace from row and column names.
Replaces consecutive spaces of length greater than one with a single space.
}

