% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorPCA.R
\name{colorPCA}
\alias{colorPCA}
\title{colorPCA}
\usage{
colorPCA(cols = NULL)
}
\arguments{
\item{cols}{a vector of valid R colors, see ?colors.}
}
\description{
Performs principle components analysis on R colors
}
\details{
Performs principle components analysis on R colors
}
\examples{
colorPCA( cols = colors()[as.integer(seq(1,657,length.out = 100))] )

}

