% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalDF.R
\name{normalDF}
\alias{normalDF}
\title{normalDF}
\usage{
normalDF(df)
}
\arguments{
\item{df}{an object of class data.frame}
}
\description{
Tests columns of objects of class data.frame for normality
}
\details{
Tests columns (characters) for normality.
If columns are factors they are converted to characters.
If columns are characters they are converted to numerics.
}
\seealso{
\code{\link[stats]{shapiro.test}}
}

