% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantMorph.R
\docType{package}
\name{PlantMorph}
\alias{PlantMorph}
\alias{PlantMorph-package}
\title{PlantMorph: A package for morphometric analysis of plant taxa.}
\description{
Provides functions for data sanitization as well as univarite and multivariate analyses.
}
\section{PlantMorph functions}{

The foo functions ...
}

