% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxa_report.R
\name{taxa_report}
\alias{taxa_report}
\title{taxa_report}
\usage{
taxa_report(chars, taxa, file = "taxa_report.Rmd", meta = TRUE)
}
\arguments{
\item{chars}{a data.frame containing samples (rows) and characters (columns).}

\item{taxa}{a data.frame containing samples (rows) and a column of class factor named 'taxa' (groupings).}

\item{file}{a character string giving the output file name.}

\item{meta}{logical specifying whether the meta data should be printed.}
}
\description{
Prepare a report for characters as a function of taxa
}
\details{
Prepare an RMarkdown report for characters.
}
\examples{
data("iris")
rownames(iris) <- paste(iris$Species, 1:nrow(iris), sep="")
colnames(iris)[5] <- "taxa"
taxa_report( chars = iris[,-5], taxa = iris[ ,5, drop=FALSE] )
}

