% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_report2.R
\name{character_report2}
\alias{character_report2}
\title{character_report2}
\usage{
character_report2(x)
}
\arguments{
\item{x}{a data.frame containing samples (rows) and characters (columns).}
}
\description{
Create a RMarkdown report about character classes and distribution
}
\details{
Prepare an RMarkdown report for characters.
}
\examples{
myData <- data.frame(matrix(1:8, ncol=4, nrow=8))
colnames(myData) <- paste('char', 1:ncol(myData), sep="_")
rownames(myData) <- paste('sample', 1:nrow(myData), sep="_")
myData[,2] <- as.character(myData[,2])
myData[,3] <- as.factor(myData[,3])
myData[,4] <- factor(myData[,4], ordered = TRUE)

\dontrun{
character_report2(myData)
}

}

